+++ src/Order/Order.php
@@ -51,6 +51,8 @@ final class Order
     private $customerTotal;
     /** @var null|string */
     private $subscriptionId;
+    /** @var null|bool */
+    private $isPaid;
 
     /**
      * @internal
@@ -82,6 +84,7 @@ final class Order
         $this->marketplaceDiscountTotal = $data['marketplaceDiscountTotal'] ?? 0.0;
         $this->customerTotal = $data['customerTotal'] ?? $data['total'];
         $this->subscriptionId = $data['subscriptionId'] ?? null;
+        $this->isPaid = \array_key_exists('is_paid', $data) ? (bool) $data['is_paid'] : null;
     }
 
     /**
@@ -228,5 +231,10 @@ final class Order
     public function getSubscriptionId(): ?string
     {
         return $this->subscriptionId;
+    }
+
+    public function isPaid(): ?bool
+    {
+        return $this->isPaid;
     }
 }