+++ CHANGELOG.md
@@ -17,8 +17,11 @@
 - Added `Wizaplace\SDK\Pim\Option\Option::getCode`
 - Added `Wizaplace\SDK\Pim\Option\OptionService::getProductOptions`
 - Added `Wizaplace\SDK\Pim\Option\OptionService::getOption`
+- Added `Wizaplace\SDK\Vendor\Order\Order::isPaid`
 - Added `Wizaplace\SDK\Vendor\Order\Order::getSubscriptionId`
+- Added `Wizaplace\SDK\Vendor\Order\OrderSummary::isPaid`
 - Added `Wizaplace\SDK\Vendor\Order\OrderSummary::getSubscriptionId`
+- Added `Wizaplace\SDK\Order\Order::isPaid`
 - Added `Wizaplace\SDK\Order\Order::getSubscriptionId`
 - Added `Wizaplace\SDK\Company\CompanyService::listSubscriptionsBy`
 - Added `Wizaplace\SDK\Subscription\SubscriptionFilter::getFilters`